update: 385x288x512 inference

+++ opensora/models/diffusion/latte/modeling_latte.py
@@ -1069,7 +1069,7 @@ class LatteT2V(ModelMixin, ConfigMixin):
                             else:  # inference may use longer video times (frames)
                                 temp_pos_embed = get_1d_sincos_pos_embed(
                                     d_size, f_size, interpolation_scale=1)  # 1152 hidden size
-                                self.temp_pos_embed = torch.from_numpy(temp_pos_embed).float().unsqueeze(0).to(hidden_states.device)
+                                self.temp_pos_embed = torch.from_numpy(temp_pos_embed).unsqueeze(0).to(hidden_states.device, dtype=hidden_states.dtype)
                                 hidden_states = hidden_states + self.temp_pos_embed
 
                         hidden_states = temp_block(
